+++ lib/src/pgn.dart
@@ -1,5 +1,4 @@
 import 'dart:math' as math;
-import 'package:fast_immutable_collections/fast_immutable_collections.dart';
 
 import 'package:meta/meta.dart';
 import './setup.dart';
@@ -493,18 +492,14 @@ class PgnEvaluation {
 @immutable
 class PgnComment {
   const PgnComment(
-      {this.text,
-      this.shapes = const IListConst([]),
-      this.clock,
-      this.emt,
-      this.eval})
+      {this.text, this.shapes = const [], this.clock, this.emt, this.eval})
       : assert(text == null || text != '');
 
   /// Comment string.
   final String? text;
 
   /// List of comment shapes.
-  final IList<PgnCommentShape> shapes;
+  final List<PgnCommentShape> shapes;
 
   /// Player's remaining time.
   final Duration? clock;
@@ -571,7 +566,7 @@ class PgnComment {
 
     return PgnComment(
         text: text.isNotEmpty ? text : null,
-        shapes: IList(shapes),
+        shapes: shapes,
         emt: emt,
         clock: clock,
         eval: eval);
@@ -601,17 +596,24 @@ class PgnComment {
 
   @override
   bool operator ==(Object other) {
-    return identical(this, other) ||
-        other is PgnComment &&
-            text == other.text &&
-            shapes == other.shapes &&
-            clock == other.clock &&
-            emt == other.emt &&
-            eval == other.eval;
+    if (identical(this, other)) return true;
+    if (other is! PgnComment) return false;
+    if (text != other.text ||
+        clock != other.clock ||
+        emt != other.emt ||
+        eval != other.eval) {
+      return false;
+    }
+    if (shapes.length != other.shapes.length) return false;
+    for (var i = 0; i < shapes.length; i++) {
+      if (shapes[i] != other.shapes[i]) return false;
+    }
+    return true;
   }
 
   @override
-  int get hashCode => Object.hash(text, shapes, clock, emt, eval);
+  int get hashCode =>
+      Object.hash(text, Object.hashAll(shapes), clock, emt, eval);
 }
 
 /// A frame used for parsing a line